New duplicate detection
This PR implements a new duplicate detection that is more alike
to the old one. It is run during prime, after all other parts
have been built. It checks the files in PRIME folder and compares
them with the corresponding one in the specifies snap folders.
If there is a correspondence, it will use filecmp.cmp() to compare
both files (first it checks the size, and if it's different, it
will say that they are different; but if the size is the same,
it will compare the contents). Only if the files are the same,
the one in PRIME is deleted.