Use the data in the snapcraft file for snap list
Currently, the developer has to specify the list of snaps to
compare with twice: one in the `build-snaps` entry, and again in
the command line of `remove_common.py`.

This patch removes the need for this, by reading the list of
snaps from the `build-snaps` entry when no snap is passed in the
command line.