Add path mapping support (#3)
* Add path mapping support

Some snaps, like gtk-common-themes, have their files in a
different layout than in the final filesystem, which prevents
`remove_common` to find some duplicated files. In the case
of gtk-common-themes, the `share` folder is in the root, when
the equivalent folder in the filesystem is at `usr/share`.

This patch allows to specify a piece of path to append, to
allow to fix this problem. For example, for gtk-common-themes,
it is enough with adding `usr` to the paths. This is achieved
by passing

  -m gtk-common-themes:usr

to the program.